[#25] Refactored functions & added rotl opcode
Some changes were made in the way these opcodes behave. Instead of
calling a get_node_at() function to fetch the second node, it removes
the need of calling another function by directly determining if stack is
up to the required length. This increases the speed of the program.

These functions were refactored: _add(), _div(), _mod(), _mul(), _sub()
and _swap.

New opcode 'rotl' was implemented with the _rotl() function.

The change was properly documented in the README.md