Ebuube/25 implement new opcodes (#28)
* [#22] Made monty accept negative numbers

Before now, monty reports an eror for negative numbers. This has been
corrected in this commit.

* [#22] Updating monty

This commit made changes that ensures monty doesn't accept invalid
strings as integers. It also makes sure that only one file is passed to
monty.

* [#25] Implement new opcodes and commenting feature

These new opcodes were implemented
* sub
* div
* mul
* mod

Added the commenting feature so that any line that begins with an hash
'#' is treated as a comment.

* [#25] Implemented the 'pchar' opcode

This involves updating the README.md to tell about such feature.

* [#25] Implement the 'pstr' opcode

This commit implements the `pstr` opcode and updated README.md with
documentation on how to use it.

* [#25] Refactored functions & added rotl opcode

Some changes were made in the way these opcodes behave. Instead of
calling a get_node_at() function to fetch the second node, it removes
the need of calling another function by directly determining if stack is
up to the required length. This increases the speed of the program.

These functions were refactored: _add(), _div(), _mod(), _mul(), _sub()
and _swap.

New opcode 'rotl' was implemented with the _rotl() function.

The change was properly documented in the README.md

* [#25] Refactored functions & added rotl opcode

Some changes were made in the way these opcodes behave. Instead of
calling a get_node_at() function to fetch the second node, it removes
the need of calling another function by directly determining if stack is
up to the required length. This increases the speed of the program.

These functions were refactored: _add(), _div(), _mod(), _mul(), _sub()
and _swap.

New opcode 'rotl' was implemented with the _rotl() function.

The change was properly documented in the README.md

* [#25] Removed previous version of files

* [#25] Implemented the 'rotr' opcode

Created the function _rotr() for this opcode and updated monty.h
and utility/get_opcode_func.c

* [#25] Implemented the 'rotr' opcode

Created the function _rotr() for this opcode and updated monty.h
and utility/get_opcode_func.c

Equally update README.md with documentation on how to use it.

* [#25] Implemented the 'stack' and 'queue' opcodes

The 'stack' opcode sets the data format to 'FIFO' which is the progams's
default state.
The 'queue' opcode sets the data format to 'LIFO'.

Updated the documentation in the README.md

* [#25] Implemented the 'stack' and 'queue' opcodes

The 'stack' opcode sets the data format to 'LIFO' which is the progams's
default state.
The 'queue' opcode sets the data format to 'FIFO'.

Updated the documentation in the README.md